UI: 포인터 기능 제거 및 수동 모드 십자선 중앙 고정, 조이스틱 제어 로직 수정

+++ routes.py
@@ -158,8 +158,8 @@ def setup_routes(app):
         """Joystick 8-direction control — enqueue large relative move then stop on release."""
         import motor_esp32 as esp
         try:
-            x = int(request.args.get('x', 0))
-            y = int(request.args.get('y', 0))
+            x = float(request.args.get('x', 0))
+            y = float(request.args.get('y', 0))
         except (TypeError, ValueError):
             return "INVALID", 400
